+++ src/tests/fp128.rs
@@ -628,6 +628,61 @@ fn saturating_add() -> Result<()> {
 }
 
 #[test]
+fn saturating_mul() -> Result<()> {
+    fn t(a: &str, b: i128, expected: &str) -> Result<()> {
+        let a: FixedPoint = a.parse()?;
+        let expected: FixedPoint = expected.parse()?;
+        assert_eq!(a.saturating_mul(b), expected);
+        assert_eq!(CheckedMul::saturating_mul(b, a), expected);
+        assert_eq!(a.cneg()?.saturating_mul(b), expected.cneg()?);
+        assert_eq!(a.saturating_mul(-b), expected.cneg()?);
+        assert_eq!(a.cneg()?.saturating_mul(-b), expected);
+        Ok(())
+    }
+
+    t("0", 0, "0")?;
+    t("3000000000000.0000000000000006", 0, "0")?;
+    t(
+        "3000000000000.0000000000000006",
+        1,
+        "3000000000000.0000000000000006",
+    )?;
+    t("-1000000000000.0000000000000002", 0, "0")?;
+    t(
+        "-1000000000000.0000000000000002",
+        3,
+        "-3000000000000.0000000000000006",
+    )?;
+    t(
+        "-1000000000000.0000000000000002",
+        -4,
+        "4000000000000.0000000000000008",
+    )?;
+    t(
+        "68603957391461.48475635294179",
+        -85204,
+        "-5845331585582084347.18029605227516",
+    )?;
+    assert_eq!(
+        fp("85550005550005550005")?.saturating_mul(85550005550005550005),
+        FixedPoint::MAX
+    );
+    assert_eq!(
+        CheckedMul::saturating_mul(14000444000, fp("14000444000.427387903")?),
+        FixedPoint::MAX
+    );
+    assert_eq!(
+        fp("-85550005550005550005")?.saturating_mul(85550005550005550005),
+        FixedPoint::MIN
+    );
+    assert_eq!(
+        CheckedMul::saturating_mul(14000444000, fp("-14000444000.427387903")?),
+        FixedPoint::MIN
+    );
+    Ok(())
+}
+
+#[test]
 fn saturating_rmul() -> Result<()> {
     fn t(a: &str, b: &str, expected: &str) -> Result<()> {
         let a: FixedPoint = a.parse()?;